refactor: do not use `@fullhuman/postcss-purgecss`
